add sum of n num code and rearrange all files in your code
