+++ C code/sum_of_n_numbers.c
