+++ python code/Python program to search an element in row-wise and column-wise sorted matrix
